redesigned buttons
