+++ button-styled-components/doc/button.mdx
@@ -1,0 +1,9 @@
+import { mdx } from '@mdx-js/react';
+import { Button } from '../index';
+import { Playground } from '@divriots/dockit-react/playground';
+import { Layout } from '~/mdx-layout-styled-components';
+export default Layout;
+
+# Button
+
+<Playground scope={{ Button }} code={`<Button>Press me</Button>`} />
