+++ button-styled-components/stories/index.stories.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { Button } from '../index';
+import { decorators } from '~/story-decorator-styled-components';
+export default { decorators }
+
+export const buttons = () => <Button>Press me</Button>;
