+++ button-theme-ui/doc/button.mdx
@@ -1,0 +1,16 @@
+import { mdx } from '@mdx-js/react';
+import { Button } from '../index';
+import { Playground } from '@divriots/dockit-react/playground';
+import { Layout } from '~/mdx-layout-theme-ui';
+export default Layout;
+
+# Button
+
+<Playground
+  scope={{ Button }}
+  code={`
+<Button mr={2}>Beep</Button>
+<Button mr={2} bg="divriots">Boop</Button>
+<Button mr={2} disabled>Disabled</Button>
+  `}
+/>
