+++ button-theme-ui/src/index.tsx
@@ -1,0 +1,1 @@
+export { Button } from 'theme-ui';
