--- button/doc/button.mdx
@@ -1,43 +1,0 @@
-import { mdx } from '@mdx-js/react';
-import {
-  Button,
-  ButtonDanger,
-  ButtonOutline,
-  ButtonPrimary,
-  ButtonClose,
-  ButtonGroup,
-  ButtonTableList,
-  ButtonInvisible,
-} from '../index';
-import { Playground } from '@divriots/dockit-react/playground';
-import { Layout } from '~/mdx-layout';
-export default Layout;
-
-# Button
-
-<Playground
-  scope={{
-    Button,
-    ButtonDanger,
-    ButtonOutline,
-    ButtonPrimary,
-    ButtonClose,
-    ButtonGroup,
-    ButtonTableList,
-    ButtonInvisible
-  }}
-  code={`
-<Button>Button</Button>
-<ButtonDanger>Button Danger</ButtonDanger>
-<ButtonOutline>Button Outline</ButtonOutline>
-<ButtonPrimary>Button Primary</ButtonPrimary>
-<ButtonInvisible>Button Invisible</ButtonInvisible>
-<ButtonClose onClick={() => window.alert('button clicked')} />
-
-<ButtonGroup display="block" my={2}>
-  <Button>Button</Button>
-  <Button>Button</Button>
-  <Button>Button</Button>
-</ButtonGroup>
-
-<ButtonTableList>Button Table List</ButtonTableList>`} />
