--- button/src/index.tsx
@@ -1,1 +1,0 @@
-export { Button, ButtonDanger, ButtonOutline, ButtonPrimary, ButtonClose, ButtonGroup, ButtonTableList, ButtonInvisible } from '@primer/components'
