+++ mdx-layout-styled-components/index.js
