+++ story-decorator-styled-components/src/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { ThemeProvider } from 'styled-components'
+
+import { theme } from '~/theme';
+
+export const decorators = [
+  (Story) => (
+    <ThemeProvider theme={theme}>
+      <Story />
+    </ThemeProvider>
+  ),
+];
